Move to singleton server / client to simplify networking. Use a fork of KryoNet.
The singleton server / client makes sense at this point because you can't reasonably have more than one server running,
nor connect to more than one server when playing.

The fork of KryoNet is because Grade kept complaining intermitently about the shaded jars used by kryonet + kryo.
They are nto neccesary as Gradle/maven manages dependencies for us. Furthermore, it hasn't been released in many
many years, so is somewhat a dormant project. This fork seems to be under some small level of active development.

+++ core/src/com/serwylo/retrowars/net/Network.kt
@@ -17,15 +17,21 @@ object Network {
         kryo.register(Client.StartGame::class.java)
     }
 
+    /**
+     * Messages sent *to* the [RetrowarsServer] sent *from* the [RetrowarsClient].
+     */
     object Server {
         class RegisterPlayer {
             override fun toString(): String = "RegisterPlayer"
         }
 
         class UnregisterPlayer
-        class UpdateScore(val score: Int) { constructor() : this(0) }
+        class UpdateScore(val score: Long) { constructor() : this(0) }
     }
 
+    /**
+     * Messages sent *to* the [RetrowarsClient] sent *from* the [RetrowarsServer].
+     */
     object Client {
 
         /**
@@ -45,7 +51,7 @@ object Network {
             override fun toString(): String = "PlayerRemoved[player id: $id]"
         }
 
-        class PlayerScored(var id: Long, var score: Int) { constructor() : this(0, 0) }
+        class PlayerScored(var id: Long, var score: Long) { constructor() : this(0, 0) }
         class StartGame
 
     }
